+++ Biflow.DataAccess/Models/Job.cs
@@ -96,7 +96,7 @@ public class Job
     [Required]
     [Display(Name = "Overtime notification limit (min, 0 = indefinite)")]
     [Range(0, 10000)]
-    public int OvertimeNotificationLimitMinutes { get; set; }
+    public double OvertimeNotificationLimitMinutes { get; set; }
 
     [Required]
     [Display(Name = "Enabled")]
