+++ Biflow.Database/biflow/Tables/Job.sql
@@ -7,7 +7,7 @@
     [UseDependencyMode]                 BIT              NOT NULL,
     [StopOnFirstError]                  BIT              CONSTRAINT [DF_Job_StopOnFirstError] DEFAULT (0) NOT NULL,
     [MaxParallelSteps]                  INT              CONSTRAINT [DF_Job_MaxParallelSteps] DEFAULT (0) NOT NULL ,
-    [OvertimeNotificationLimitMinutes]  INT              CONSTRAINT [DF_Job_OvertimeNotificationLimitMinutes] DEFAULT(0) NOT NULL,
+    [OvertimeNotificationLimitMinutes]  FLOAT            CONSTRAINT [DF_Job_OvertimeNotificationLimitMinutes] DEFAULT(0) NOT NULL,
     [IsEnabled]                         BIT              CONSTRAINT [DF_Job_IsEnabled] DEFAULT (1) NOT NULL,
     [CreatedBy]                         NVARCHAR(250)    NULL,
     [LastModifiedBy]                    NVARCHAR(250)    NULL,
